requirements.txt
aws-cdk.core
aws-cdk.aws-iam
aws-cdk.aws-lambda
aws-cdk.aws-cloudformation
aws-cdk.aws-codedeploy
aws-cdk.aws-codebuild 
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline_actions
aws-cdk.pipelines 